+++ beacon-chain/blockchain/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "chain_info.go",
         "error.go",
         "execution_engine.go",
+        "forkchoice_chain_info.go",
         "forkchoice_update_execution.go",
         "head.go",
         "head_sync_committee_info.go",
